Move isLoggedIn to selector
